requirements.txt
selenium==4.15.0
pytest==7.4.3
pytest-html==4.1.1
allure-pytest==2.13.2
webdriver-manager==4.0.1